# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a  server BackChat.",
        "title": "BACKCHAT Api",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:9888",
    "basePath": "/api/v1",
    "paths": {
        "/example/signup": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "writes the user to the database",
                "parameters": [
                    {
                        "description": "user",
                        "name": "SignUpInput",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.AuthorizationUser"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/migrations.Users"
                        }
                    }
                }
            }
        },
        "/example/userID": {
            "get": {
                "description": "getting all users",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/migrations.Users"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controllers.AuthorizationUser": {
            "type": "object",
            "required": [
                "email",
                "password",
                "username"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "username": {
                    "description": "ID_user           uint   `json:\"id_user\" binding:\"required\"`",
                    "type": "string"
                },
                "verification_code": {
                    "type": "string"
                }
            }
        },
        "migrations.Users": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "id_user": {
                    "type": "integer"
                },
                "password": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "username addin": {
                    "type": "string"
                },
                "verification_code": {
                    "type": "string"
                }
            }
        }
    }
}